move file_size from dataframe.io to utils
